test: add API coverage for Sanitizer singleton methods
These are the methods used by Rails. Also move these methods into the
Sanitizer class definition.
